move classes and set up phpcs
