+++ composer.json
@@ -8,11 +8,13 @@
 	},
 	"require-dev": {
 		"phpunit/phpunit": "^9.0",
-		"automattic/jetpack-autoloader": "^2"
+		"automattic/jetpack-autoloader": "^2",
+		"woocommerce/woocommerce-sniffs": "^1.0"
 	},
 	"config": {
 		"allow-plugins": {
-			"automattic/jetpack-autoloader": true
+			"automattic/jetpack-autoloader": true,
+			"dealerdirect/phpcodesniffer-composer-installer": true
 		}
 	}
 }
