+++ phpcs.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0"?>
+<ruleset name="WooCommerce Coding Standards">
+	<description>My projects ruleset.</description>
+	
+	<!-- Configs -->
+	<config name="minimum_supported_wp_version" value="4.7" />
+	<config name="testVersion" value="7.4-" />
+
+	<!-- Rules -->
+	<rule ref="WooCommerce-Core" />
+
+	<rule ref="WordPress.WP.I18n">
+		<properties>
+			<property name="text_domain" type="array" value="snapchat_for_woocommerce" />
+		</properties>
+	</rule>
+
+	<rule ref="PHPCompatibility">
+		<exclude-pattern>tests/</exclude-pattern>
+		<exclude-pattern>vendor/</exclude-pattern>
+		<exclude-pattern>node_modules/</exclude-pattern>
+	</rule>
+</ruleset>
